music-player-backend-new/src/main/java/dev/georgi/musicplayerbackendnew/controller/SearchController.java
package dev.georgi.musicplayerbackendnew.controller;

import com.fasterxml.jackson.databind.JsonNode;
import dev.georgi.musicplayerbackendnew.service.SearchService;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1")
public class SearchController {
    private final SearchService searchService;

    public SearchController(SearchService searchService) {
        this.searchService = searchService;
    }

    @GetMapping("/search/{searchTerm}")
    public JsonNode searchSongsArtists(@PathVariable String searchTerm) {
        return searchService.searchSongsArtists(searchTerm);
    }
}
